Enforce column casing in SQLCheckExtractors
Uses the parent extractor's _is_uppercase_names property to
determine if the column should be upper cased in the
SQLColumnCheckExtractor's _get_input_facets() method. These
methods have also been moved from util to the extractors
directly as they were not utils to be used across extractors.
